Added markers to separate json file
