+++ src/scss/global.scss
@@ -1,6 +1,5 @@
 button {
   color: #B0B0B0;
-  display: flex;
   background: #555;
   padding: 10px 15px;
   border: 0;
