+++ src/scss/menu.scss
@@ -2,16 +2,15 @@
   position: fixed;
   top: 0;
   width: 250px;
-  height: 96%;
-  margin: 0.5em;
+  top: 0.5em;
+  left: 0.5em;
+  bottom: 0.5em;
   padding: 0.5em;
   transition: transform ease-in-out 250ms;
   &.left {
     z-index: 10;
     background: #4f0000;
-    left: 0;
     transform: translateX(-300px);
-    margin-bottom: 2em;
   }
   &.visible.left {
     transform: translateX(0);
@@ -44,6 +43,8 @@
   height: 12em;
   padding: 1em;
   color: white;
+  box-shadow: inset 0 0 35px #000000;
+  border: 1px solid #000;
 }
 .SideMenu-logo {
   height: 200px;
